requirement.txt
#Añadir librerias

pandas
plotly-express
streamlit
